Refactor file format: create tag app
