+++ book_lovers/book_lovers/books/api/serializers.py
@@ -1,5 +1,6 @@
 from rest_framework import serializers
-from books.models import Book, Publisher
+
+from book_lovers.books import Book, Publisher
 
 
 class PublisherSerializer(serializers.ModelSerializer):
