+++ book_lovers/book_lovers/books/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class BooksConfig(AppConfig):
-    name = 'books'
+    name = 'book_lovers.books'
