+++ book_lovers/book_lovers/books/models.py
@@ -1,5 +1,5 @@
 from django.db import models
-
+from django.conf import settings
 
 class Book(models.Model):
