+++ book_lovers/book_lovers/books/urls.py
@@ -1,6 +1,7 @@
-from books import views
-from django.contrib.auth import views as auth_views
 from django.conf.urls import url
+from django.contrib.auth import views as auth_views
+
+from book_lovers.books import views
 
 urlpatterns = [
     url(r'^$', views.BooksListView.as_view(), name='list'),
