+++ book_lovers/book_lovers/books/views.py
@@ -1,10 +1,10 @@
-from django.shortcuts import render
 from django.contrib import messages
 from django.contrib.auth.mixins import LoginRequiredMixin
-from django.views.generic import CreateView, DetailView, UpdateView, ListView, DeleteView
-from django.urls import reverse
 from django.db.models import Q
-from books.models import Book,Publisher
+from django.urls import reverse
+from django.views.generic import CreateView, DetailView, UpdateView, ListView, DeleteView
+
+from book_lovers.books import Book
 
 
 class BooksActionMixin(object):
