+++ book_lovers/book_lovers/my_books/admin.py
@@ -1,9 +1,5 @@
 from django.contrib import admin
-from .models import Tag,Profile
+from .models import Profile
 
 
-
-
-
-admin.site.register(Tag)
 admin.site.register(Profile)
