+++ book_lovers/book_lovers/my_books/api_urls.py
@@ -1,7 +1,8 @@
-from rest_framework import routers
-from my_books import api_views
 from django.conf.urls import url
+from rest_framework import routers
 from rest_framework_swagger.views import get_swagger_view
+
+from book_lovers.my_books import api_views
 
 schema_view = get_swagger_view(title='Pastebin API')
