+++ book_lovers/book_lovers/my_books/models.py
@@ -7,11 +7,7 @@ from django.conf import settings
 
 
 
-class Tag(models.Model):
-    name = models.CharField(max_length=50)
-    
-    def __str__(self):
-        return self.name
+
 
 
 class Profile(models.Model):
