+++ book_lovers/book_lovers/tags/admin.py
@@ -1,0 +1,4 @@
+from django.contrib import admin
+from .models import Tag
+
+admin.site.register(Tag)
