+++ book_lovers/book_lovers/tags/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class TagsConfig(AppConfig):
+    name = 'book_lovers.tags'
