+++ book_lovers/book_lovers/tags/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from .models import Tag
+
+
+class TagSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Tag
+        fields = '__all__'
