+++ book_lovers/book_lovers/tags/urls.py
@@ -1,0 +1,11 @@
+# from rest_framework import routers
+# from book_lovers.tags import views
+#
+# router = routers.DefaultRouter()
+# router.register(r'tags', views.TagViewSet)
+#
+# urlpatterns = [
+#
+# ]
+#
+# urlpatterns += router.urls
