+++ book_lovers/book_lovers/tags/views.py
@@ -1,0 +1,9 @@
+# from rest_framework import viewsets, permissions
+# from .serializers import TagSerializer
+# from .models import Tag
+#
+#
+# class TagViewSet(viewsets.ModelViewSet):
+#     queryset = Tag.objects.all()
+#     serializer_class = TagSerializer
+#     permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
