+++ book_lovers/book_lovers/templates/books/book_detail.html
@@ -1,4 +1,4 @@
-{% extends 'my_books/base.html' %}
+{% extends 'my_books/../my_books/base.html' %}
 
 
 {% block content %}
